requirements.txt
fastapi
uvicorn
playwright
google-genai
google-auth
pydantic
python-dotenv
jinja2
streamlit
